refactor: move asset placeholders to a dedicated file
relocates image-related constants and placeholders from the general constants utility to a dedicated assets file for better separation of concerns and improved code organization.
