+++ projects/client/src/lib/sections/toast/_internal/ToastEpisodeCover.svelte
@@ -5,7 +5,7 @@
   import { useEpisodeSpoilerImage } from "$lib/features/spoilers/useEpisodeSpoilerImage";
   import type { EpisodeEntry } from "$lib/requests/models/EpisodeEntry";
   import type { ShowEntry } from "$lib/requests/models/ShowEntry";
-  import { EPISODE_COVER_PLACEHOLDER } from "$lib/utils/constants";
+  import { EPISODE_COVER_PLACEHOLDER } from "$lib/utils/assets";
   import { episodeActivityTitle } from "$lib/utils/intl/episodeActivityTitle";
   import { UrlBuilder } from "$lib/utils/url/UrlBuilder";
