+++ projects/client/src/routes/branding/+page.svelte
@@ -2,7 +2,7 @@
   import * as m from "$lib/features/i18n/messages.ts";
   import Branding from "$lib/sections/branding/Branding.svelte";
   import TraktPage from "$lib/sections/layout/TraktPage.svelte";
-  import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_COVER } from "$lib/utils/assets";
 </script>
 
 <TraktPage
