+++ projects/client/src/routes/discover/+page.svelte
@@ -12,7 +12,7 @@
   import RecommendedList from "$lib/sections/lists/recommended/RecommendedList.svelte";
   import TrendingList from "$lib/sections/lists/trending/TrendingList.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
-  import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/assets";
 
   const { mode: type, useSeasonalFilters } = useDiscover();
   const { themeFilters } = useSeasonalTheme();
