+++ projects/client/src/routes/lists/official/[list]/+page.svelte
@@ -3,7 +3,7 @@
   import TraktPage from "$lib/sections/layout/TraktPage.svelte";
   import TraktPageCoverSetter from "$lib/sections/layout/TraktPageCoverSetter.svelte";
   import UserListPaginatedList from "$lib/sections/lists/user/UserListPaginatedList.svelte";
-  import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_COVER } from "$lib/utils/assets";
   import { mapToMediaType } from "../../../users/[user]/lists/[list]/_internal/mapToMediaType";
   import type { PageProps } from "../[list]/$types";
   import { useListSummary } from "./useListSummary";
