+++ projects/client/src/routes/lists/smart/create/+page.svelte
@@ -6,7 +6,7 @@
   import TraktPageCoverSetter from "$lib/sections/layout/TraktPageCoverSetter.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
   import SmartListCreator from "$lib/sections/smart-lists/SmartListCreator.svelte";
-  import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_COVER } from "$lib/utils/assets";
 
   const { mode } = useDiscover();
