+++ projects/client/src/routes/media/popular/+page.svelte
@@ -5,7 +5,7 @@
   import PopularPaginatedList from "$lib/sections/lists/popular/PopularPaginatedList.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
 
-  import { DEFAULT_SHARE_MOVIE_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_MOVIE_COVER } from "$lib/utils/assets";
 </script>
 
 <TraktPage
