+++ projects/client/src/routes/search/+page.svelte
@@ -12,7 +12,7 @@
   import TraktPage from "$lib/sections/layout/TraktPage.svelte";
   import TraktPageCoverSetter from "$lib/sections/layout/TraktPageCoverSetter.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
-  import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_COVER } from "$lib/utils/assets";
   import { isMobileAppleDevice } from "$lib/utils/devices/isMobileAppleDevice";
 
   const query = $derived(page.url.searchParams.get("q")?.trim());
