+++ projects/client/src/routes/shows/popular/+page.svelte
@@ -5,7 +5,7 @@
 
   import PopularPaginatedList from "$lib/sections/lists/popular/PopularPaginatedList.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
-  import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/assets";
 </script>
 
 <TraktPage
