+++ projects/client/src/routes/shows/recommended/+page.svelte
@@ -5,7 +5,7 @@
 
   import RecommendedPaginatedList from "$lib/sections/lists/recommended/RecommendedPaginatedList.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
-  import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/assets";
 </script>
 
 <TraktPage
