+++ projects/client/src/routes/social/activity/+page.svelte
@@ -9,7 +9,7 @@
 
   import ActivityPaginatedList from "$lib/sections/lists/activity/ActivityPaginatedList.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
-  import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_COVER } from "$lib/utils/assets";
 
   const { mode } = useDiscover();
