+++ projects/client/src/routes/users/[user]/start-watching/+page.svelte
@@ -6,7 +6,7 @@
   import TraktPageCoverSetter from "$lib/sections/layout/TraktPageCoverSetter.svelte";
   import UpNextPaginatedList from "$lib/sections/lists/progress/UpNextPaginatedList.svelte";
   import NavbarStateSetter from "$lib/sections/navbar/NavbarStateSetter.svelte";
-  import { DEFAULT_SHARE_COVER } from "$lib/utils/constants";
+  import { DEFAULT_SHARE_COVER } from "$lib/utils/assets";
 </script>
 
 <TraktPage
